Cube changed to a Toblerone chocolate
Changed the shape into a triangular prism by changing size. Commented out .cullBlackfaces(); , so that the chocolate stays black and doesn't go transparent. I changed the values in the main function to change background color hue(?) and to have the object completely black because it is a bitter chocolate. And lastly I changed the constant multiplied with time to change the rotation speed on x and y axises.