All the space balls now rotate and they have better color with no distortion
Added rotation to all objects. Changed the view distance to fit all of the movement in. Added uniform float time under vertex shader to use it in main function to create holes on the objects by editing the vertex shader.  Commented out cullblackfaces and uncommented the depthtest functions to have better colors on the objects without distortion.