Images added, 2 Demos added
I made 2 different projects with different styles and functions. The cool one is like the name says, cooler. The other one has my classic Uganda Knuckles texture. They both have ambient light sources rotating and movement of the camera and the main object. Added floor function of tan funciton of texture to make a cool texture animation. Added skybox to both files as well. Changed the html so that it contains the files.